solve reading sequence problem and use unsigned instead of size_t

+++ lib/store.c
@@ -4,7 +4,7 @@
 #include <string.h>
 
 void fread_string(char **string, FILE *file) {
-	size_t length;
+	unsigned length;
 	fread(&length, sizeof(length), 1, file);
 	*string = malloc(sizeof(char) * length);
 	fread(*string, sizeof(char), length, file);
@@ -14,7 +14,7 @@ void fwrite_string(char *string, FILE *file) {
 	if (!string) {
 		return;
 	}
-	size_t length = strlen(string) + 1;
+	unsigned length = strlen(string) + 1;
 	fwrite(&length, sizeof(length), 1, file);
 	fwrite(string, sizeof(char), length, file);
 }
